Chats loading issue
The issue was with the encryption and decryption of the chats being send and recieved. The _encrypt method in the chat_textfield_bloc and the _decrypt method in the message_text page.

The CHAT_SECRET key from the .env file was unable to process the HEX being sent when a person sends a message.

Please do review the chat_secret i added relative to one that was there. Please note. The code that was being used before for this is just uncommented out and not deleted.